pyproject.toml
[project]
name = "nanoarray-python-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "google-search-results>=2.4.2",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "litellm>=1.81.16",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
